+++ src/format/utf8.h
@@ -1,4 +1,4 @@
-unsigned int utf8_decode(char *s, long *u);
-unsigned int utf8_encode(long *u, char *s);
+unsigned int utf8_decode(char *s, int *u);
+unsigned int utf8_encode(int *u, char *s);
 unsigned int utf8_isfull(char *s, int b);
 unsigned int utf8_size(char *s);
